add method put checking email and quey param with age

+++ app.js
@@ -41,15 +41,33 @@ const server = http.createServer(async (req, res) => {
 
 
         writeFiles(res, 200, JSON.stringify(usersName), res)
+    } else if (req.url.match(/^\/api\/users\?age=([0-9a-zA-Z]+)$/) && req.method === 'GET') {
+        const users = await JSON.parse(await readFile('db', 'users.json'))
+        const query = path.basename(req.url).split('=');
+        const age = query[1]
+        let sortedAge = [...users]
+        if (age === 'min') {
+            sortedAge = sortedAge.toSorted((a,b) => a.age - b.age)
+        } else if (age === 'max') {
+            sortedAge = sortedAge.toSorted((a,b) => b.age - a.age)
+        } else {
+            writeFiles(res, 404, JSON.stringify({message:'Not allow'}),'application/json')
+        }
+        writeFiles(res,200,JSON.stringify(sortedAge),'application/json')
     } else if (req.url === '/api/users' && req.method === "POST") {
         req.on('data', async (chunk) => {
             let body = JSON.parse(chunk.toString());
             body.id = Date.now()
             const users = await JSON.parse(await readFile('db', 'users.json'))
-            users.push(body)
-            fs.unlink(path.join(__dirname, 'db', 'users.json'), (err) => { console.log(err) });
-            fs.appendFile(path.join(__dirname, 'db', 'users.json'), JSON.stringify(users));
-            res.end(JSON.stringify(users))
+            const user = users.find((e) => e.email === body.email)
+            if (user) {
+                writeFiles(res, 200, JSON.stringify({message:`Email -> ${body.email} is already exist`}),'application/json')
+            } else {
+                users.push(body)
+                fs.unlink(path.join(__dirname, 'db', 'users.json'), (err) => { console.log(err) });
+                fs.appendFile(path.join(__dirname, 'db', 'users.json'), JSON.stringify(users));
+                res.end(JSON.stringify(users))
+            }
         });
     } else if (req.url.match(/^\/api\/users\/([0-9]+)$/) && req.method === "PATCH") {
         const id = path.basename(req.url);
@@ -62,6 +80,21 @@ const server = http.createServer(async (req, res) => {
                 return;
             }
             findedUser.name = body.name
+            fs.unlink(path.join(__dirname, 'db', 'users.json'), (err) => { console.log(err) });
+            fs.appendFile(path.join(__dirname, 'db', 'users.json'), JSON.stringify(users))
+            res.end(JSON.stringify(users))
+        })
+    } else if (req.url.match(/^\/api\/users\/([0-9]+)$/) && req.method === "PUT") {
+        const id = path.basename(req.url);
+        req.on('data', async (chunk) => {
+            let body = JSON.parse(chunk.toString());
+            const users = await JSON.parse(await readFile('db', 'users.json'))
+            const findedUser = users.find((e) => e.id === id)
+            if (!findedUser) {
+                error404(res)
+                return;
+            }
+            Object.assign(findedUser,body)
             fs.unlink(path.join(__dirname, 'db', 'users.json'), (err) => { console.log(err) });
             fs.appendFile(path.join(__dirname, 'db', 'users.json'), JSON.stringify(users))
             res.end(JSON.stringify(users))
